Refactor UTF-8 input handling: use code points rather than UTF-16 units
Discard bytes that don't decode as UTF-8.

Stop calling VkKeyScan on the two halves of a surrogate pair.  It's
impossible for VkKeyScan to return a useful result.

+++ src/agent/InputMap.h
@@ -34,7 +34,7 @@ class InputMap {
 public:
     struct Key {
         uint16_t virtualKey;
-        uint16_t unicodeChar;
+        uint32_t unicodeChar;
         uint16_t keyState;
 
         std::string toString() const;
